+++ pix_sweeping_driver/src/vcu2_acu_chassis_err_code3.cc
@@ -23,7 +23,7 @@ void Vcu2AcuChassisErrCode3::Parse() {
 }
 
 
-// config detail: {'bit': 56, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rr_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 56, 'description': '右后电控故障2', 'enum': {1: 'MCU_CHIP_ERROR', 2: 'PRE_CHARGE_ERROR', 4: 'MOSTFET_FAULT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rr_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassisrredserrorcode2() {
   Byte t0(*(bytes + 7));
   int32_t x = t0.get_byte(0, 1);
@@ -33,11 +33,11 @@ int Vcu2AcuChassisErrCode3::chassisrredserrorcode2() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 40, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rf_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 40, 'description': '右前电控故障2', 'enum': {1: 'MCU_CHIP_ERROR', 2: 'PRE_CHARGE_ERROR', 4: 'MOSTFET_FAULT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rf_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassisrfedserrorcode2() {
   Byte t0(*(bytes + 5));
   int32_t x = t0.get_byte(0, 1);
@@ -47,11 +47,11 @@ int Vcu2AcuChassisErrCode3::chassisrfedserrorcode2() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 24, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lr_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 24, 'description': '左后电控故障2', 'enum': {1: 'MCU_CHIP_ERROR', 2: 'PRE_CHARGE_ERROR', 4: 'MOSTFET_FAULT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lr_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassislredserrorcode2() {
   Byte t0(*(bytes + 3));
   int32_t x = t0.get_byte(0, 1);
@@ -61,11 +61,11 @@ int Vcu2AcuChassisErrCode3::chassislredserrorcode2() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 8, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lf_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 8, 'description': '左前电控故障2', 'enum': {1: 'MCU_CHIP_ERROR', 2: 'PRE_CHARGE_ERROR', 4: 'MOSTFET_FAULT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lf_eds_error_code2', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassislfedserrorcode2() {
   Byte t0(*(bytes + 1));
   int32_t x = t0.get_byte(0, 1);
@@ -75,11 +75,11 @@ int Vcu2AcuChassisErrCode3::chassislfedserrorcode2() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 48, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rr_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 48, 'description': '右后电控故障', 'enum': {1: 'OVER_CURRENT', 2: 'UNDER_VOLTAGE', 4: 'HALL_SENSOR_ERROR', 8: 'OVER_VOLTAGE', 16: 'MOTOR_BLOCK', 32: 'SPEED_FEEDBACK', 64: 'MCU_EEPROM_WRITE_ERROR', 128: 'OVERHEAT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rr_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassisrredserrorcode() {
   Byte t0(*(bytes + 6));
   int32_t x = t0.get_byte(0, 1);
@@ -89,11 +89,11 @@ int Vcu2AcuChassisErrCode3::chassisrredserrorcode() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 32, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rf_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 32, 'description': '右前电控故障', 'enum': {1: 'OVER_CURRENT', 2: 'UNDER_VOLTAGE', 4: 'HALL_SENSOR_ERROR', 8: 'OVER_VOLTAGE', 16: 'MOTOR_BLOCK', 32: 'SPEED_FEEDBACK', 64: 'MCU_EEPROM_WRITE_ERROR', 128: 'OVERHEAT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_rf_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassisrfedserrorcode() {
   Byte t0(*(bytes + 4));
   int32_t x = t0.get_byte(0, 1);
@@ -103,11 +103,11 @@ int Vcu2AcuChassisErrCode3::chassisrfedserrorcode() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 16, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lr_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 16, 'description': '左后电控故障', 'enum': {1: 'OVER_CURRENT', 2: 'UNDER_VOLTAGE', 4: 'HALL_SENSOR_ERROR', 8: 'OVER_VOLTAGE', 16: 'MOTOR_BLOCK', 32: 'SPEED_FEEDBACK', 64: 'MCU_EEPROM_WRITE_ERROR', 128: 'OVERHEAT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lr_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassislredserrorcode() {
   Byte t0(*(bytes + 2));
   int32_t x = t0.get_byte(0, 1);
@@ -117,11 +117,11 @@ int Vcu2AcuChassisErrCode3::chassislredserrorcode() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
 
-// config detail: {'bit': 0, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lf_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'int'}
+// config detail: {'bit': 0, 'description': '左前电控故障', 'enum': {1: 'OVER_CURRENT', 2: 'UNDER_VOLTAGE', 4: 'HALL_SENSOR_ERROR', 8: 'OVER_VOLTAGE', 16: 'MOTOR_BLOCK', 32: 'SPEED_FEEDBACK', 64: 'MCU_EEPROM_WRITE_ERROR', 128: 'OVERHEAT'}, 'is_signed_var': False, 'len': 8, 'name': 'chassis_lf_eds_error_code', 'offset': 0.0, 'order': 'motorola', 'physical_range': '[0|255]', 'physical_unit': '', 'precision': 1.0, 'type': 'enum'}
 int Vcu2AcuChassisErrCode3::chassislfedserrorcode() {
   Byte t0(*(bytes + 0));
   int32_t x = t0.get_byte(0, 1);
@@ -131,7 +131,7 @@ int Vcu2AcuChassisErrCode3::chassislfedserrorcode() {
   x <<= 7;
   x |= t;
 
-  int ret = x;
+  int ret =  static_cast<int>(x);
   return ret;
 }
